integration o full scrn and HUD to latest work
everything feels slow probably because the screen is larger lol

+++ src/mechanics/skills/CastFireBall.java
@@ -1,7 +1,7 @@
 package mechanics.skills;
 
 import entity.core.complex.ComplexUnit;
-import entity.core.element.ProjectileEmitter;
+import entity.core.particle.ProjectileEmitter;
 import entity.player.FireBall;
 import mechanics.skills.core.Skill;
 
@@ -16,7 +16,7 @@ public class CastFireBall extends Skill{
 	@Override
 	public void use(ComplexUnit u) {
 		if(u.getEnergy() >= this.getCost()){
-			u.setEnergy(u.getEnergy() - this.getCost());
+			u.setEnergy((float)(u.getEnergy() - this.getCost()));
 			FireBall fb = new FireBall(u.getTileMap(), u.isFacingRight(), u.getState());
 			fb.setPosition(u.getX(), u.getY());
 			ProjectileEmitter.getInstance().emit(fb);
